+++ src/routes/models.js
@@ -11,9 +11,11 @@ const ProjectSchema = require('../models/projects');
  */
 router.get('/:projectId', auth, async (req, res) => {
     try {
+        // Collect models from DB and return them
         models = await MlModelSchema.find({ project_id: req.params.projectId });
         res.json(models);
     } catch(err) {
+        // On failure log the error message and return status 500
         console.log(err.message);
         res.status(500).json({ 
             message: "Server Error"
@@ -25,10 +27,23 @@ router.get('/:projectId', auth, async (req, res) => {
  * POST method to create new models
  */
 router.post('/:projectId/new', auth, async (req, res) => {
+    // Collect project from DB to verify if the user owns this project
+    let project = await ProjectSchema.findOne({ _id: req.params.projectId, user_id: req.user.id });
+
+    // If not return error
+    if(!project) {
+        return res.status(404).json({
+            message: "Incorrect Project chosen"
+        });
+    }
+
+    // Get name and parameters of dataset
     const { name, parameters } = req.body;
     try {
+        // Make model object
         mlmodel = new MlModelSchema({
             name: name,
+            user_id: req.user.id,
             project_id: req.params.projectId,
             parameters: {
                 batch_size: parameters.batch_size,
@@ -38,16 +53,19 @@ router.post('/:projectId/new', auth, async (req, res) => {
 
         mlmodel._id = mongoose.Types.ObjectId();
 
+        // Save the object into DB
         await mlmodel.save();
 
-        let project = await ProjectSchema.findOne({ _id: req.params.projectId })
-
+        // Push the model object into mlModels array 
         project.mlModels.push(mlmodel);
 
+        // Save project
         await project.save();
 
+        // Return model object in JSON format
         res.json(mlmodel);
     } catch(err) {
+        // On failure log the error message and return status 500
         console.log(err.message);
         res.status(500).json({
             message: "Server Error"
@@ -60,23 +78,39 @@ router.post('/:projectId/new', auth, async (req, res) => {
  */
 router.delete('/:projectId/remove', auth, async (req, res) => {
     try {
-        const output = await MlModelSchema.deleteOne({ name: req.body.name });
-        if(!output) {
+        // Collect project from DB to verify if the user owns this project
+        let project = await ProjectSchema.findOne({ _id: req.params.projectId, user_id: req.user.id });
+
+        // If not return error
+        if(!project) {
+            return res.status(404).json({
+                message: "User not authorized"
+            });
+        }
+
+        // Find ML Model with the specified name and in the same project and delete it
+        const output = await MlModelSchema.deleteOne({ name: req.body.name, project_id: req.params.projectId, user_id: req.user.id });
+        // If no elements were deleted, return status 404
+        if(!output.deletedCount) {
             return res.status(404).json({
                 message: "Model does not exist"
             });
         }
 
-        let project = await ProjectSchema.findOne({ _id: req.params.projectId })
-
+        // Find ML Model in project and delete it
         let index = project.mlModels.findIndex(x => x.name === req.body.name);
 
         project.mlModels.splice(index, 1);
 
+        // Save project
         await project.save();
 
-        res.json(req.body.name + " model deleted"); 
+        // Return status 200
+        return res.status(200).json({
+            message:  req.body.name  + " deleted"
+        }); 
     } catch(err) {
+        // On failure log the error message and return status 500
         console.log(err.message);
         return res.status(500).json({
             message: "Server Error"
@@ -88,27 +122,46 @@ router.delete('/:projectId/remove', auth, async (req, res) => {
  * PATCH method to update name of the model
  */
 router.patch('/:projectId/update', auth, async(req, res) => {
+    // Collect the new name from request body
     const { newName } = req.body;
     try {
-        const newModel = await MlModelSchema.findOneAndUpdate({ name: req.body.name }, {$set: {name: newName}});
+        // Collect project from DB to verify if the user owns this project
+        let project = await ProjectSchema.findOne({ _id: req.params.projectId, user_id: req.user.id });
+
+        // If not return error
+        if(!project) {
+            return res.status(404).json({
+                message: "User not authorized"
+            });
+        }
+
+        // Find ML Model with the specified name and in the same project and update its name
+        const newModel = await MlModelSchema.findOneAndUpdate({ name: req.body.name, user_id: req.user.id }, {$set: {name: newName}});
+        // If no elements were changed, return status 404
         if(!newModel) {
             return res.status(404).json({
                 message: "Model does not exist"
             });
         }
 
+        // Save the changes to DB
         await newModel.save();
 
-        let project = await ProjectSchema.findOne({ _id: req.params.projectId })
-
+        // Find the Ml Model in project
         let index = project.mlModels.findIndex(x => x.name === req.body.name);
 
+        // Change the name
         project.mlModels[index].name = newName;
 
+        // Save the project
         await project.save();
 
-        res.json("Model name updated"); 
+        // Return status 200
+        return res.status(200).json({
+            message:  req.body.name  + " updated"
+        }); 
     } catch(err) {
+        // On failure log the error message and return status 500
         console.log(err.message);
         return res.status(500).json({
             message: "Server Error"
